move utils to anomaly detection directory
